used lu instead of zu

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,21 +9,21 @@
  */
 size_t print_list(const list_t *h)
 {
-    size_t s = 0;
+	size_t s = 0;
 
-    while (h)
-    {
-        if (h->str == NULL)
-        {
-            printf("[0] (nil)\n");
-        }
-        else
-        {
-            printf("[%zu] %s\n", h->len, h->str);
-        }
-        h = h->next;
-        s++;
-    }
+	while (h)
+	{
+		if (h->str == NULL)
+		{
+			printf("[0] (nil)\n");
+		}
+		else
+		{
+			printf("[%lu] %s\n", (unsigned long)h->len, h->str);
+		}
+		h = h->next;
+		s++;
+	}
 
-    return (s);
+	return (s);
 }
